backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic==2.10.3
python-multipart==0.0.18
httpx==0.28.1
pillow==11.1.0
PyMuPDF==1.25.3
python-dotenv==1.0.1
aiofiles==24.1.0
